feat(home page): Added a home page to app
